[Email] Split N001_Email.ino into N001_data_struct
